Port the original dashboard card repros consolidation
